更改包目录结构
