+++ app/src/main/java/tty/balanceyourio/MainActivity.kt
@@ -8,7 +8,7 @@ import android.support.v4.view.ViewPager
 import android.view.Menu
 import android.view.MenuItem
 import kotlinx.android.synthetic.main.activity_main.*
-import tty.util.MainFragmentAdapter
+import tty.balanceyourio.adapter.MainFragmentAdapter
 
 
 class MainActivity : AppCompatActivity(), BottomNavigationView.OnNavigationItemSelectedListener, ViewPager.OnPageChangeListener {
@@ -38,7 +38,7 @@ class MainActivity : AppCompatActivity(), BottomNavigationView.OnNavigationItemS
 
     }
 
-    private var adapter:MainFragmentAdapter?=null
+    private var adapter: MainFragmentAdapter?=null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
