+++ app/src/main/java/tty/balanceyourio/adapter/MainFragmentAdapter.kt
@@ -1,4 +1,4 @@
-package tty.util
+package tty.balanceyourio.adapter
 
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
